requirements.txt
# requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
slack-bolt==1.18.0
slack-sdk==3.26.1
langchain==0.1.5
langchain-openai==0.0.5
langchain-community==0.0.19
python-dotenv==1.0.0
pinecone-client==2.2.4
pydantic==2.5.2
pydantic-settings==2.1.0
google-cloud-firestore==2.13.1
openai==1.3.5